Simplify management of traits file.
