+++ google/cloud/bigtable/bigtable_client.bzl
@@ -45,7 +45,6 @@ bigtable_client_hdrs = [
     "internal/async_retry_unary_rpc.h",
     "internal/async_retry_unary_rpc_and_poll.h",
     "internal/bulk_mutator.h",
-    "internal/bytes_traits.h",
     "internal/client_options_defaults.h",
     "internal/completion_queue_impl.h",
     "internal/common_client.h",
@@ -92,7 +91,6 @@ bigtable_client_srcs = [
     "instance_update_config.cc",
     "internal/async_bulk_apply.cc",
     "internal/bulk_mutator.cc",
-    "internal/bytes_traits.cc",
     "internal/completion_queue_impl.cc",
     "internal/common_client.cc",
     "internal/google_bytes_traits.cc",
