+++ google/cloud/bigtable/bigtable_client_unit_tests.bzl
@@ -42,6 +42,7 @@ bigtable_client_unit_tests = [
     "internal/async_retry_multi_page_test.cc",
     "internal/async_retry_unary_rpc_test.cc",
     "internal/bulk_mutator_test.cc",
+    "internal/google_bytes_traits_test.cc",
     "internal/grpc_error_delegate_test.cc",
     "internal/prefix_range_end_test.cc",
     "mutation_batcher_test.cc",
